\name{FactoMineR-package}
\alias{FactoMineR-package}
\alias{FactoMineR}
\docType{package}
\title{Multivariate Exploratory Data Analysis and Data Mining with R
}
\description{
The method proposed in this package are exploratory mutlivariate methods such as principal component analysis, correspondence analysis or clustering.
}
\details{
\tabular{ll}{
Package: \tab FactoMineR\cr
Type: \tab Package\cr
Version: \tab 1.34\cr
Date: \tab 2014-09-26\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}

\author{
Francois Husson, Julie Josse, Sebastien Le, Jeremy Mazet

Maintainer: \email{husson@agrocampus-ouest.fr}
}

\references{
Le, S., Josse, J. & Husson, F. (2008). FactoMineR: An R Package for Multivariate Analysis. Journal of Statistical Software. 25(1). pp. 1-18.
\url{https://www.jstatsoft.org/v25/i01/}

A website: \url{http://factominer.free.fr/}

Some videos: \url{https://www.youtube.com/playlist?list=PLnZgp6epRBbTsZEFXi_p6W48HhNyqwxIu}
}
\keyword{Factor analysis}

