\name{print.spMCA}

\alias{print.spMCA}

\title{Print the specific Multiple Correspondance Analysis (spMCA) results}

\description{
Print the specific Multiple Correspondance Analysis (spMCA) results.
}

\usage{
\method{print}{spMCA}(x, file = NULL, sep = ";", \dots)
}

\arguments{
  \item{x}{an object of class spMCA}
  \item{file}{A connection, or a character string naming the file to print to.  If NULL (the default), the results are not printed in a file}
  \item{sep}{character string to insert between the objects to print (if the argument file is not NULL}
  \item{\dots}{further arguments passed to or from other methods}
}

\author{Francois Husson \email{Francois.Husson@agrocampus-ouest.fr}}

\seealso{ \code{\link{spMCA}}, \code{\link{write.infile}}}

\keyword{print}
