\name{print.AFDM}

\alias{print.AFDM}

\title{Print the Multiple Factor Analysis of mixt Data (AFDM) results}

\description{
Print the Multiple Factor Analysis of mixt Data (AFDM) results.
}

\usage{
\method{print}{AFDM}(x, file = NULL, sep = ";", \dots)
}

\arguments{
  \item{x}{an object of class AFDM}
  \item{file}{A connection, or a character string naming the file to print to.  If NULL (the default), the results are not printed in a file}
  \item{sep}{character string to insert between the objects to print (if the argument file is not NULL}
  \item{\dots}{further arguments passed to or from other methods}
}

\author{Jeremy Mazet, Francois Husson \email{Francois.Husson@agrocampus-ouest.fr}}

\seealso{ \code{\link{AFDM}}}

\keyword{print}
